refactor(connections): extract StoredConnection Codable model into its own file
